Renamed redirect files
